requirements.txt
pyeda~=0.28.0
xmldict~=0.4.1
matplotlib~=3.4.2
networkx~=2.5.1
click~=8.0.1
pandas~=1.3.1
lxml~=4.6.3
loguru~=0.5.3
antlr4-python3-runtime
pyahocorasick~=1.4.2
